requirements.txt
# Core dependencies compatible with Python 3.13
langchain>=0.2.0
langchain-openai>=0.3.0
langchain-community>=0.1.0
chromadb>=0.5.0
openai>=1.100.0
tiktoken>=0.12.0
pydantic>=2.0.0
fastapi>=0.110.0
uvicorn>=0.29.0
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.1.0
numpy>=2.1.0  # Required for Python 3.13
pandas>=2.2.0
pyautogen>=0.2.26
sentence-transformers>=2.6.0
weaviate-client>=4.6.0
redis>=5.0.0
pyyaml>=6.0.0
jinja2>=3.1.0
markupsafe>=2.1.0
python-multipart>=0.0.9
python-slugify>=8.0.0
fake-useragent>=1.5.0
ddgs>=5.3.0  # duckduckgo-search 的新版本名称
arxiv>=2.1.0
wikipedia>=1.4.0
tavily-python>=0.3.0
llmsherpa>=0.1.0
unstructured>=0.13.0
pdfplumber>=0.10.0
langchain-chroma>=1.0.0
aiohttp>=3.9.0
# Ensure setuptools is up to date for Python 3.13 compatibility
setuptools>=80.0.0
wheel>=0.45.0